requirements.txt
selenium
python-dotenv
streamlit
psycopg2-binary
pytest